started work on test config, moved mock to a new file for organization
